mtd: spi-nor: refactor read op creation to make a template op
commit 72c6d526c7d067d7af55164a697867ec932cff6e from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

A template of the read op will be needed in a upcoming commit. So,
refactor the code to create a read op in spi_nor_read_data() to a
separate function that returns the template of the op. The caller can
then fill in the details like address, data length, and the data buffer.
Update spi_nor_read_data() to use this template.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>