mmc: sdhci_am654: Add workaround for card detect debounce timer
commit 59bb7a4deacc537f52df281324f3e7dfee6eff87 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

There is a one time delay because of a card detect debounce timer in the
controller IP. This timer runs as soon as power is applied to the module
regardless of whether a card is present or not and any writes to
SDHCI_POWER_ON will return 0 before it expires. This timeout has been
measured to be about 1 second in am654x and j721e.

Write-and-read-back in a loop on SDHCI_POWER_ON for a maximum of
1.5 seconds to make sure that the controller actually powers on.

Signed-off-by: Faiz Abbas <faiz_abbas@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>