Revert "arm64: dts: ti: k3-am65-main: add timesync_router node"
commit ff46c1d6dce2f67d39b5b53ab35001f6d2fdf5c4 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

This reverts commit 3f2ac41102af4488c047b6db97f4c0ef4cef510e.

The TimeSync Interrupt Router module is currently represented as
a pinctrl device in Linux, with the individual pinmux settings
configuring the interrupt router connections. This IP is an
Interrupt Router and is in general firewalled away and needs
to be programmed using System Firmware.

Remove this pinctrl device in preparation to rework and model
it as a proper TI-SCI Interrupt Router device.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>