mtd: spi-nor: default to addr_width of 3 for configurable widths
commit 8a5dfc668b5558491bce17d01298b32f8325664f from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit f9acd7fa80be6ee14aecdc54429f2a48e56224e8 upstream

JESD216D.01 says that when the address width can be 3 or 4, it defaults
to 3 and enters 4-byte mode when given the appropriate command. So, when
we see a configurable width, default to 3 and let flash that default to
4 change it in a post-bfpt fixup.

This fixes SMPT parsing for flashes with configurable address width. If
the SMPT descriptor advertises variable address width, we use
nor->addr_width as the address width. But since it was not set to any
value from the SFDP table, the read command uses an address width of 0,
resulting in an incorrect read being issued.

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>
Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>