clk: Correctly annotate 's32gen1_clocks_init' function
commit fc1f055dc9e67f25e7ba0cb96cffe239e841df51 from
https://source.codeaurora.org/external/autobsps32/linux

Because 's32gen1_clocks_init' references other functions
annotated with '__init' macro, 's32gen1_clocks_init' should be
also annotated with '__init'.

This way, both the caller and the callee will end up in the same
section, avoiding the situation when a function that is in a
long-lasting section refers to data in one of the discardable
sections.

Issue: ALB-5413
Signed-off-by: Ciprian Marian Costea <ciprianmarian.costea@nxp.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>