s32g274a: ddr: Add workaround for Erratum ERR050543
commit dfb914654bef47b16b628f1605a2d03ee15dc725 from
https://source.codeaurora.org/external/autobsps32/linux

This commit adds a new driver based on Linux Kernel Timers.

The Linux driver operates in two scenarios:

1. Booting at high temp (> 85 C):
U-boot applies ERR050543 workaround. In this case, the
Linux driver periodically checks if ERR050543 workaround
is no longer needed (with a time period of 1 sec).
If it is not, refresh rate is updated and auto derating
is turned on (ERR050543 changes are partially removed).
After this operation, no periodically polling of MR4 register
is being performed by the driver.

2. Booting at low/room temp (< 85 C):
U-boot does not apply the ERR050543 workaround. Therefore, no
polling is being performed from Linux at any point in time,
in the future.

Issue: ALB-5557
Signed-off-by: Ciprian Marian Costea <ciprianmarian.costea@nxp.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>