net: phy: DP83822: Fix WoL in config init to be disabled
commit e50d86a7ba4bc5195bef3ca9f852638aecd372a1 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit 600ac36b5327c949f6754be106e2a08e5d81e3a0 upstream.

The WoL feature should be disabled when config_init is called and the
feature should turned on or off  when set_wol is called.

In addition updated the calls to modify the registers to use the set_bit
and clear_bit function calls.

Fixes: 3b427751a9d0 ("net: phy: DP83822 initial driver submission")
Signed-off-by: Dan Murphy <dmurphy@ti.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>