net: phy: DP83822: Add setting the fixed internal delay
commit 48ce3448f0dcc82c987956c37017bf8caa709e6f from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit 8095295292b564583bb5bee5ba69924c2fd40434 upstream.

The DP83822 can be configured to use the RGMII interface. There are
independent fixed 3.5ns clock shift (aka internal delay) for the TX and RX
paths. This allow either one to be set if the MII interface is RGMII and
the value is set in the firmware node.

Signed-off-by: Dan Murphy <dmurphy@ti.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>