arm64: dts: ti: k3-j7200-som-p0: Set initial TX start to 18
commit 99ba16ea03dd142e92da74ff13416491addf4497 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

TX = 16 is on the border of the range of valid TX values for this board
with the Cypress S28 flash. As a result, the PHY tuning sometimes fails
to find rxlow at that point, causing the algorithm to iterate over the
TX range taking a hit on the tuning time.

TX = 18 is much more stable in comparison. Tuning consistently succeeds
with it. Use it as the initial value to stop the PHY tuning from taking
longer.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>