net: ethernet: ti: icss_iep: Allow requesting ts capabilities before interface is up
commit f1119daea74999ce182aa8958831a61c95063ec7 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

iep driver registers ptp_clock in icss_iep_init(), which gets called
only duing ndo_open(). This sequence does not allow to query ts capabilities
when ndo interface is not up. But it is expected to be able to query
timestamp capabilities even the interface is not up. So, register ptp
clock and populate other timestamping capabilities in icss_iep_get(),
which gets called in probe of ndo device.

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>